@@ -1,34 +1,15 @@
-import Utils from './core';
+import { BaseClass, Utils } from './core';
 
-export class BaseModel {
-    constructor(data: any = null) {
-        if (data !== null) {
-            for (let p of data) {
-                if (this[p] !== undefined) {
-                    let v = data[p];
-                    switch (typeof (this[p])) {
-                        case 'boolean':
-                            this[p] = Utils.parseBoolean(v);
-                            break;
-                        case 'number':
-                            if (Utils.isNullOrEmpty(v)) {
-                                this[p] = 0;
-                            } else {
-                                let v_ = 0;
-                                if (v.toString().split('.').length > 1) {
-                                    v_ = parseFloat(v);
-                                } else {
-                                    v_ = parseInt(v);
-                                }
-                                this[p] = isNaN(v_) ? 0 : v_;
-                            }
-                            break;
-                        default:
-                            this[p] = v;
-                            break;
-                    }
-                }
-            }
-        }
+export class BaseModel extends BaseClass {
+    constructor(data?: any) {
+        super();
+
+        Promise.resolve(null).then(() => {
+            this.parseObject(data);
+        });
+    }
+
+    public parseObject(data: any): void {
+        Utils.parseObject(data, this);
     }
 }
