@@ -2,10 +2,11 @@ import { BaseModel } from '../core/base.model';
 import { CameraType } from './enums/CameraType';
 
 export class Camera extends BaseModel {
-    id: number = 0;
-    type: CameraType = CameraType.NONE;
-    roverId: number = 0;
-    fullName: string = '';
+
+    public cameraId: number = 0;
+    public fullName: string = '';
+    public roverId: number = 0;
+    public type: CameraType = CameraType.NONE;
 
     constructor(data: any = null) {
         super(data);
